Extract ChangeHistory as an object
This change makes ChangeHistory a first-class
object that exposes some methods for adding and
updating items. This means the app doesn’t need
to know the details of how the change history is
represented.

Other than refactoring, the driver for this change
is to remove all reference of ‘redrafted’ from the
code base. Previously, change history was so
tangled up with ‘redrafted’ that it makes sense to
fix this first before continuing with that work.