Add DFID peer review dropdown
Although not currently used as a facet because the label "unreviewed"
makes no sense to researchers, work needs to be done in both
finder-frontend and specialist-frontend to support a piece of
metadata that is either present or absent.

In the meantime, this lets DFID record the peer reviewed status of an
output so that this information is not lost and when the works to
finder-frontend and specialist-frontend are done, the information
will be available to face the public.

In order to do this and keep it out of the finder we need hard-coded
values in a helper.