Discard erroneous drafts from Publishing API
There are a few content items in Publishing API where the
content_id mismatches with the documents in the old
Specialist Publisher application. I have confirmed they are
all in a draft state and can be discarded. We need to do
this so that we can republish these documents from V1 as
part of migrating to the rebuild app. Otherwise, these
documents conflict on base_path in Publshing API.