Improve inline attachment matching…
This is more-or-less a rewrite of how inline
attachment snippets are matched against
attachments on the document.

This commit also introduces a rake task for
reporting on invalid attachments snippets in the
bodies of documents which can be used to feedback
to editors about incorrect snippets.

The report could also be used to surface
information to users in the app, potentially.