Specify publishing_app when accessing content items
Due to the change in the Publishing API you could now view documents
that are associated with a different application. The risk of this is
very low as they would need to have the same document_type but I've
added the app specification anyway so it doesn't bite someone later.