Restore attachments when a document is published
When a document has attachments and has previously been unpublished,
the attachments will be in a soft-deleted state and unavailable in
the asset api. If the document is redrafted and published we should
restore these attachments on the asset api. This can be done in the
background as the availability of the assets doesn't immediately
impact the publishing application.