Add response errors and warnings to Document
Whether to publish or save a document relies not only on
its validity when initialized from a form but also the response
obtained from the publishing api. Warnings do no prevent a
document from being saved but may prevent it being published.
Certain errors where a base path is already in use need to be
propagated up to the view so these are also stored on the document.