Use YAML instead of JSON for the finder configuration
I think YAML is better here because it allows whitespace to improve
readability and it allows comments.

It also makes clear that the file here is not the finder content item,
but rather a kind of configuration file. Making this clearer will allow
us to move more behaviour into this file, rather than the class.