Use Rails.root to remove relative resolution of path
Without using Rails.root this will based off of the working directory of
the application. The reason this hasn't presented any problems before is
because this is nearly always the value of the Rails root.

However if we utilise the rails server feature of running this as a
daemon the value of the working directory is "/" which causes this path
to be looked for in the wrong place.