Publish UTAAC and Tax Tribunal decisions
This is a rake task that will persist until we've got MOJ live.  The documents
referred to in the CSV files has been pre-checked and approved for publication.

We're publishing the items listed in the CSV files so that we can make the
UTAAC and Tax Tribunal finders live later on.

We'll remove the CSVs from the repo when we're done.  We have subsequent
finders to go live so we'll retain the rake task/worker for a bit longer.