Test finder schemas match specialist document schema
Move the FinderSchemaConverter class from govuk-content-schemas.
We used to have a script there that generated the `metadata` part
of the `specialist_document` content schema, based on the finder
schemas published by specialist publisher:
https://github.com/alphagov/govuk-content-schemas/blob/66b11b4004b49064b6aae16715158740977dc706/bin/convert_finder_schemas

This was a bit out of date, and didn't work with all the schemas we've
created. We've changed this to a set of tests in specialist publisher,
so that consistency is always checked for new schemas, unless they are
explicitly added to the `IGNORED_FINDERS` array.

Trello: https://trello.com/c/lFZnNhs4/349-create-a-service-assessment-finder