Add assessment_date for Service Standard Report
We are adding `assessment_date` as part of the metadata sent to
Publishing API. This means that editors will be able to set the date of
a given assessment.

This is a follow up work on moving all the service assessments and self
certification documents away from GDS Data blog
(https://gdsdata.blog.gov.uk/all-service-assessments-and-self-certification/)
and move them to Specialist Publisher.

Trello:
https://trello.com/c/ADPnDXsE/385-add-assessment-date-to-service-standard-specialist-document

Paired with: @klssmith