Test length of subscription list names
This commit amend the test for the length of subscription list names to take into account the change of approach in how the names are generated. It now checks to make sure either the prefix (where only the prefix is used) or the plural prefix with all topics appended is a maximum of 255 characters long. Any finders that are not live are excluded from this test to allow for pre-production testing.