Make hidden_indexable_content indexable
All MOJ finders use a hidden_indexable_field which should be indexed by
Rummger but currently isn't.  This PR adds code that includes the field
in a document's indexable content, when the field is present.

NOTE: This functionality used to work so likely it is a regression
introduced when specialist publisher was rebuilt.