Add a task to republish documents to rummager
Currently we have in place a task to republish documents to Publishing API and
Rummager, but in some cases we only want to republish to rummager in
order to update the document.

This commit adds an extra task into the republish.rake file to republish
all documents to rummager.

Trello:
https://trello.com/c/iMotNdUv/461-add-more-document-type-to-search