Make rummager finder task consistent with Publishing API task
This commit makes the publishing-api finder loader more generic so that it can also be used by the rummager finder publishing task. It also makes the rummager finder publishing task more consistent with its publishing-api equivalent, and changes it to send to rummager immediately instead of placing changes in a queue.

This is what the publishing-api task does, and since these tasks are called manually, it makes sense for their results to be immediately available.