Add “select all”/“clear all” buttons to dropdown menus
This commit adds “select all”/“clear all” buttons to dropdown menus. These buttons allow a user to quickly select or clear all options in a dropdown menu with one action, which is useful if there are many options that are often all selected.

The buttons require JavaScript and are also added using JavaScript, so they won’t appear if the browser has JavaScript disabled.