Revert "Add rake task to find pages with missing search content"
This reverts commit 8ed2f1c9d69a3ce50e8a6666970a297b108b0677.

The temporary rake task has been run in production and most of the
missing search content has been fixed.

The remaining missing content can't be fixed just now because the
RepublishWorker will only send content to rummager if the latest version
of a document is in the `published` state, but some specialist content
has extra `draft` versions.

We have decided not to handle that case now. Once search publishing has
been moved to the publishing API, we will work out how to handle content
with `draft` versions.