Add AllDocumentsFinder.find_each
This talks to the publishing api and fetches every page of results for a
document type so that callers can iterate over all the documents without
having to handle pagination themselves.  We use `find_each` for this to
mirror the function from AR that does something similar by batching up
records to avoid large queries.