Allow document types to say if they are exportable
For now only business finance support schemes are exportable, but we can
extend it to others as we need to.  We only do it for business finance
support scheme at the moment because we're building a report that used to
exist for these in publisher when they were managed in there.