Always return humanized keys from FinderSchema#humanized_facet_name
If the supplied key does not have a facet definition we would return the
key as provided.  It's unlikely this would actually be a "humanized" name
which means when using `Document#humanized_attributes` the hash keys are a
mix of actual humanized names from the facet definitions and underscored
symbols.  Ultimately the only place `humanized_attributes` is used is in
the view where we have to call `.to_s.humanize` on the supposedly
humanized attribute name just in case.

To avoid this mixed return from `Document#humanized_attributes` we move
the `.to_s.humanize` into `FinderSchema#humanized_facet_name`, although we
only do it to the fallback value when it's not present in the facet
definitions (on the assumption that if it is in the facet definitions
it'll be a humanized string already).