Introduce presenters for exporting documents to CSV
We have DocumentExportPresenter which acts as a factory so you can get a
presenter for a given document class, and then we have the individual
document presenters that know how to represent each document type as CSV.

For now only the BusinessFinanceSupportScheme document type is actually
exportable, so it's the only one that has a presenter.  The contents of
the CSV generated here are as close as possible to the version that used
to be generated by publisher (see it's final form when it was removed
in [this commit][last-publisher-version]).  Although the CSV contents
are inspired by publisher's version, we take inspiration for the shape
of the presenter from the whitehall CSV generators.

[last-publisher-version]: https://github.com/alphagov/publisher/pull/607/commits/e9ef795c4afab16f7d53e84289d18c6b3884efe1#diff-4face7cd83286e29589b48da13937b0e