Send public_updated_at to email-alert-api
From https://github.com/alphagov/specialist-publisher/pull/1061 we are now sending the published version of a document as retrieved from the publishing-api to email-alert-api. The side effect is that the change note is no longer sent since specialist-publisher ignores it when retrieving documents.

This commit reverts to sending the original document but with `public_updated_at` added in from the published document, which was the original issue, before sending it to email-alert-api. It also reverts changes made in https://github.com/alphagov/specialist-publisher/pull/1062 and https://github.com/alphagov/specialist-publisher/pull/1064 which were unsuccessful and are no longer required.