Add rake task to unpublish unsupported schemes
Northern Ireland and Wales entries within business support finder are not
supported by the business support finder team, so we need to remove them.  This
adds a temporary rake task to unpublish these schemes in bulk.  We also
temporarily add a CSV list of the schemes that need to be removed.