Don't add "XXX All" buttons to single selection select2 dropdowns
If the list of things to select from is long we want to use select2 to
make using the list easier, but it doesn't make sense to add the "Select
all" and "Clear all" convienence buttons to the UI if the dropdown is
used to select a single item.  So we restrict the behaviour enhancement to
select2 instances that have the `multiple` attribute.