Re-point drug safety updates to MHRA account
This commit changes the drug safety update finder to point to the MHRA GovDelivery account for signing up to email alerts. It also allows this field to make its way to the content item, where it will be picked up by finder-frontend, and removes the signup page text and content ID since this page will no longer exist.