Rename Document#organisations to Document#schema_organisations
Document#organisations returns the static orgs from the
schema file. In order to use the links organisations effectively
we need to rename this so #organisations can be used for links.

This commit also adds a permissions check for '<document_type>_editor'
permissions so that permissions can be spread across organisations
for a specific document type."