Add rake task to edit a document's base_bath
Trello card:
https://trello.com/c/kTKWF48j/530-document-how-to-change-a-url-in-specialist-publisher

How to change a URL in Specialist Publisher wasn't a documented process.
Making it into a rake task seems a nicer solution than simply
documenting how to do it and then having to do it manually.

Now, in order to edit a document's base_path, it's possible to just use
the rake task passing the document's `content_id` and the new
`base_path` as arguments.