Bugfix for minor Drug Safety Updates being published as major updates
This fixes a bug where a document would be sent to publishing_api
as a major update even if the publisher user had set the
update_type to minor.

The bug would occur because first_published_at could be sent as
nil when the user clicks publish (probably due to PR #1362).

Because first_published_at was nil, we would consider the document
to be a first draft, and set the update_type to major before
sending the document to the publishing_api.

This was reported as a bug via 2nd line at:
https://govuk.zendesk.com/agent/tickets/3603919

The fix copies the DfidResearchOutput approach,
which overrides first_draft? to use the state
history rather than the first_published_at date.

Rather than overriding the first_draft? method
this replaces the Document class method as its
better for consistency.

DfidResearchOutput and DrugSafetyUpdate are the only
classes that extend Document and also provide
a way to set a custom first_published_at,
therefore there isnt a need for further changes. But
because it is quite easy to introduce this bug we
thought it would be worth changing the Document class.