Fix casing of country names for UK MCABs
Attempting to create a document which applies to all countries fails
with errors about Jamaica, Lebanon, and the Ivory Coast.  Jamaica and
Lebanon are failing because they should be lowercase (like all the
other countries), Ivory Coast is failing because it's missing from the
schema (which will be fixed in govuk-content-schemas).