Remove redirect URL check when unpublishing
This removes the check to ensure that when unpublishing a document
the redirect is pointing to live content. This because the publishing
api currently doesn't support lookups on base path of multi-part
content. Hence publishers are currently unable to redirect to multipart
content such as guides. This behaviour is also inconsistent with other
publishing applications.