Remove DocumentListExportRequest model
The model is used for two things:

1. determining the document type
2. checking if the export has been processed yet

(1) is used for permissions.  This can be solved by putting the type
in the URL *and* in the S3 filename.  It needs to go in both so that a
user can't just change their request URL to fetch exports of arbitrary
document types.

(2) is used to serve a 404 if it's not done yet.  But there's an
easier way to do that: just serve a 404 if the file doesn't exist in
S3.  It's good to do that anyway.