Stop sending hidden_indexable_content to Email Alert API
This change is prompted by seeing rather lengthy prose appearing in the
Email Alert API database that seems rather out of place. It transpires
the Specialist Publisher declares all of the fields flagged as metadata
for a format as tags, even those that are lengthy prose.

I've taken the rather lazy approach here of solving this with an
exceptional condition in the EmailAlertPresenter.

Pondering what the right way to solve this is somewhat riddled with
rabbit holes. It seems somewhat dubious that this presenter class holds
the logic as to what is a tag and what isn't, this feels more something
that is in the scope of the model. However, once pondering that the
question arises about why is Specialist Publisher not telling the
Publishing API what the tags are which indicates the path forward is
probably to be re-structuring the Publishing API details to have a tags
field and to consider whether the metadata hash is needed or not.