Move lib files to app/lib
In Rails 4.0 the lib directory was remove from the auto/eager load paths
because it resulted in a lot of junk being eager loaded in production
without the developer often being aware of this. The recommend pattern
now is to add an app/lib directory for non-specific application code.

The services.rb file is left in place because this is explicitly
required from the Sidekiq workers and the republish.rake file and
so doesn't need to be auto/eager loaded like the other files.