Normalise the require statements for services.rb
Since rake tasks and when running the application locally doesn't eager
load the source code there was a chance of hitting a missing constant
error depending on what url or task was chosen. Fix this by requiring
the services.rb file in all the places that it is used and removing
some unnecessary requires that services.rb loads anyway.