Use params for constructing the edit form URL
When the locale was introduced in to the URL, as the locale can be
changed, this means that if you fill in the form, including changing
the locale, but there are validation errors, the form will be
re-rendered using the values you entered including the changed locale
in the URL, but this change hasn't been sent to the Publishing API
yet. The result is that Specialist Publisher can't find the document
when you submit the form.

To avoid this, use the URL parameters to construct the URL, as these
shouldn't change half way through handling the request.