Use a select box for the locale
This is better, as it's clearer to the user what the options are. It
would be better to have more understandable names to go along with the
locale codes, but I'm not sure where to get those from. Also, maybe
the list of locales should be fetched from the schema rather than be
copied in to the code.