Specialist Publisher - Bulk import of Protected food and Drink names
This PR introduce a new Rake task of which responsiblity is to deal with bulk imports.

One import have been developed to bulk-import Protected Food and Drink names from a CSV file, of which examples were provided.

The import usage has been documented in the Rake task description.
When the process is run, it will provide useful messages regarding any document that could not be uploaded.
The code uses a `-ImportService` which is responsible to get the CSV::Row (or an hash of information) and save a document from them. This service is also responsible to construct useful error messages.
There is also a `-Parser` object which is used to parse the content of the Row data into the set of attributes that the ProtectedFoodDrinkName model expects.
The Rake task is responsible for parsing the CSV correctly, send each row in turn to the ImportService and report any error to stdout.

Test coverage is present, whith a rake task test that function as full integration test for the feature.

ff