Publishing in Bulk via Rake Task
This PR contains the code that implements a publishing in bulk feature.
The feature is accessible via Rake task. I used some code found in the `republish` task functionality. However, in the case of `publishing`, the code looks simpler as it does not have to validate states and states versions.
It will only work with documents that are currently in a Draft status.
This tasks also allow to run the publish all functionality on a subset of document types rather than all existing documents in `draft` mode.
Discussions has to happen still about the background job or in-line publiation. Looking at existing functionality, seems like a background job is the way to go.