Ensure Greek characters are transliterated
At the moment there's an issue where Greek characters aren't been
transliterated by ActiveSupport, and instead we end up with an empty
string:

```
> "Στερεά Ελλάδα".parameterize
=> ""
```

This is an issue because this output is used the generate the slug of
documents. After this change we get a more usable output:

```
> "Στερεά Ελλάδα".parameterize
=> "sterea-ellada"
```

I got the letters from https://gist.github.com/Achillefs/925267 and I'm
not sure about the validity of them.

This has come up as part of the "Protected Food & Drink Names" finder
where some titles are composed entirely of Greek letters and aren't able
to be saved correctly because the slug ends up being an empty string.