Add registered_name field
This is a new field for the protected food & drink names finder that the
department would like to use. At the moment this is currently held in
the title of the documents, but it turns out there are some duplicate
entries, so they want the "Registered Name" to be the same for those
documents and have different titles.