Add email filters to flood reports
Add ability to receive emails on particular categories of flood and
coastal erosion reports. I've updated a few lines of code from Becca's
original commit.

* I've changed `required` to `true` which forces users to select a
  category when they're on the email sign-up page (they can select all
  categories too) - see an [example] page where this is enabled. This
  seems reasonable as it makes sure users make an active choice about
  what they want to receive.

* I've added `signup_content_id` which was missing (can't verify this is
  unique due to not having VPN access).

`email_filter_by` was mentioned in the card as being potentially
deprecated so may be not be needed, I assume because of [the TODO
comment]. However `email_filter_by` is still used by Finder Frontend to
determine whether the email facet selection can be [modified or not]
when signing up for emails.

[example]: https://www.gov.uk/european-structural-investment-funds/email-signup
[the TODO comment]: https://github.com/alphagov/specialist-publisher/blob/1ea8b91c3d3fbe66560dcb573d0ee4dc719de54c/app/presenters/finder_signup_content_item_presenter.rb#L75
[modified or not]: https://github.com/alphagov/finder-frontend/blob/bc6dbbc959a9f0c18401354d0f8614a6650ac34c/app/presenters/subscriber_list_params_presenter.rb#L36-L38