Update key to be more descriptive
The previous key name was a bit vague and as these values get passed
around different apps it'd be good to be more specific. The driver for
this change was that we need to add a tag to the [allowed list] in Email
Alert API which allows a SubscriberList to be created for the facets.

[allowed list]: https://github.com/alphagov/email-alert-api/blob/master/lib/valid_tags.rb#L2