Replace master with main
The `master` branch has been renamed to `main`.

"Both Conservancy and the Git project are aware that the initial branch name,
‘master’, is offensive to some people and we empathize with those hurt by the
 use of that term".
[Software Freedom Conservancy](https://sfconservancy.org/news/2020/jun/23/gitbranchname/).

This PR updates the URLs for repos with a default branch of `main`. If these URLS were left
as they were before then Github would cleverly redirect them. However, this reduces our use of
the term and will help to keep naming consistent going forward.