Add internal notes to UKMCAB documents
When unpublishing a document, publishing api accepts an "explanation" attribute [1].
We can use this to populate the internal note field when we build a UKMCAB document.

Ideally this new field would be added to the FORMAT_SPECIFIC_FIELDS defined in the UKMCAB,
rather than bleeding into the parent Document class, but FORMAT_SPECIFIC_FIELDS are populated with the
the contents of the metadata hash of the content item. Which is not where the explanation key lives. Amending
the widely used document_builder class to do something special for a UKMCAB's didn't seem like the best idea.

Government frontend will not display this new field when it renders this document type as the key "internal_notes" is
excluded [2]

[1] https://github.com/alphagov/gds-api-adapters/blob/2c17390a1a9dab648adef124bd605d95e12be460/lib/gds_api/publishing_api.rb#L186
[2] https://github.com/alphagov/government-frontend/blob/7e20c048a37479dad95e9252788ccdc192f8ac75/app/presenters/specialist_document_presenter.rb#L201