Ensure date labels link to the associated field
At the moment, the fieldset label 'Date of occurrence' and the
'day', 'month', 'year' labels are not associated with the
corresponding inputs. This means that screen readers don't acknowledge
them. It also means that errors are not linked.

This commit ensures there are no orphaned labels when the date_fields
partial is employed.

It also removes placeholders from the inputs as these are bad
practice. Clear labelling is preferable.