Increase test coverage of attachments_controller.rb
Some of the `else` statements of the file were not being tested.
This commit increases the test coverage associated with the file
by testing these and tests for the presence of flashes.

This commit also reformats some of the longer lines of code
and restructures the tests into context blocks.