Stop using rake test helper for bulk_imports rake spec
The `rake_task: true` argument was causing the rake tasks to be reloaded
which meant code coverage was lost on them. This task isn't needed as we
load the rake tasks as part of rails_helper.rb and we can use the
reenable method on the task to set them up for test runs. This
subsequently deletes the spec helper file that was used to set up this
action, since nothing uses it anymore.

This also corrects some naming in the file.