Stop using rake test helper in bulk_imports_spec
The `rake_task: true` argument was causing the rake tasks to be reloaded
which meant code coverage of other rake tasks was lost. This task isn't needed
becasue we load the rake tasks in rails_helper and can use the reenable method
on the tasks to set them up for test runs. This subsequently makes the
load_rake_tasks helper file redundant, meaning it can be deleted.

This also corrects some naming in the file.