Set up the Jasmine testing framework
Used jasmine-browser-runner to setup the framework. This is required in
order to add the necessary JS tests for enabling CD for Specialist
Publisher. Rake tasks added and the Rakefile amended so that JS tests
run with every rake. Setup was done in accordance with the steps
outlined in the "Migrate apps to jasmine-browser-runner" ticket.