Authenticate users against doc types their org can access
The check on root path checks for users with "editor" permissions checks that the user's organisation can access the first finder in the list of schemas. This schema is AAIB reports and not all users will be able to access them.
Rather than checking against the first document model, check against the first model that the user's org can access.