Add licence transaction schema
Adds MVP features to create the new SP licence.

Users will be able to filter the licence they're looking for by country,
sector and activity. I've also included this in the metadata of each
result. We may want to remove the metadata if licences are tagged to
multiple sectors / activities.

Haven't included email functionality as this feature is not yet required.

The sector / activity lists are dummy values for testing. Real values
will be confirmed at a later date.