Update sector values in licence transaction schema
JSON was generated by running:

```
 bundle exec rake licence_transaction:import_industry_sectors
```

`JSON.dump` outputs the JSON file on to one line which is not very
readable. The editor's (in this case VS Code) "format document" function
has been used to prettify the code, which is why the formatting in the
rest of the schema has changed.