Adds config option for licences
We need to mark licences as `imported` to prevent Specialist Publisher
assuming we're creating a first draft, which has a few knock on effects,
e.g first draft change note and generating a base_path from the title.

An `imported` method could be added to the model but this would effect
new licences created via Specialist Publisher. Adding a flag to the
params and then pulling it out seems the best option.