Adds class to import licences
This class is temporary and just being used to populdate the finder with
licences for User Research purposes. If / when we actually import all
the licences for real, we'll need to update this script in a number of
ways (take into account drafts, decide links etc).

We only want to import licences identified by ids in the
`common_licence_identifiers.txt` file.  This is explained in a previous
commit.

Currently, we only want published licences as we're using the imported
licences in User Research and thus don't want to test documents that are
not public. We need to save the new licences and then publish them as we don't
support draft finders.

There is validation on each new licence which ensures data integrity
(title present etc).

Industry / location metadata will be imported in a seperate piece of
work.

We're keeping all links for each licence document until it's decided
what is being persisted to the new licence format. Due to that reason we
can't use the baked in method `new_licence.save` as
`DocumentLinksPresenter` only supports a hardcoded list of links.

I've decided to use my own Publishing API response for a licence after
an API call, rather than using `GovukSchemas` to load an example as it
ensures the response is realistic.