Skip importing licences that already exist
We've added a feature to skip importing a licence if it has already been
imported in a previous migration. This is to ensure we can rerun the
task, without issue, if there is an error midway through which causes
only some licences to be imported. Errors may include validation errors
or things like Publishing API being unavailable / communication errors.

To handle this situation we could've tackled it in three different ways,
e.g:

1) Skip already imported licence

2) Unpublish old licence and publish new licence

3) Update old licence and publish

I didn't choose option 2 as we don't expect the licence to have been
updated / we don't really care as the migration is for User Research and
data will be reset anyway. When we run this for real, we expect the
publishing app (Publisher) to be locked and so licences won't be
updatable at that point mitigating the risk.

I also didn't choose option 3 as this would require multiple editions to be
published / change history that we don't want or need, as it would be
better to have a fresh edition on import.

`base_path` seems a good candidate for identifying the imported licence
as it will be unique to each licence (can't use content_id as it is
generated on the fly).